+++ src/parsecmd.go
@@ -3,6 +3,7 @@ package main
 /*
 #include <stddef.h>
 extern unsigned int dword_5d4594_823684;
+extern int nox_cheat_allowall;
 
 void nox_xxx_consoleTokenAddPair_4444C0(wchar_t* tok, wchar_t* tok2);
 int nox_gui_console_Print_450B90(unsigned char typ, wchar_t* str);
@@ -110,6 +111,10 @@ type consolePrinter struct{}
 
 func (consolePrinter) Printf(cl parsecmd.Color, format string, args ...interface{}) {
 	consolePrintf(cl, format, args...)
+}
+
+func cheatAllowAll(v bool) {
+	C.nox_cheat_allowall = C.int(bool2int(v))
 }
 
 var parseCmd = parsecmd.NewConsole(consolePrinter{}, strMan)